+++ pkg/krt/validations_general.go
@@ -1,4 +1,4 @@
-package main
+package krt
 
 import (
 	"regexp"
