+++ pkg/krt/validations_process.go
@@ -1,4 +1,4 @@
-package main
+package krt
 
 import (
 	"fmt"
